Added new atomist url to config Small fixes to polish the generic prod request implementation Implemented the automation handler for the generic prod requests.
