+++ nucleus/src/main/java/za/co/absa/subatomic/adapter/project/rest/ProjectResourceBase.java
@@ -14,4 +14,6 @@ public class ProjectResourceBase extends ResourceSupport {
     protected String name;
 
     protected String description;
+
+    protected String owningTenant;
 }
