+++ nucleus/src/main/java/za/co/absa/subatomic/adapter/project/rest/ProjectResourceBaseAssembler.java
@@ -1,6 +1,7 @@
 package za.co.absa.subatomic.adapter.project.rest;
 
 import org.springframework.hateoas.mvc.ResourceAssemblerSupport;
+
 import za.co.absa.subatomic.infrastructure.project.view.jpa.ProjectEntity;
 
 public class ProjectResourceBaseAssembler extends
@@ -18,6 +19,7 @@ public class ProjectResourceBaseAssembler extends
             resource.setProjectId(entity.getProjectId());
             resource.setDescription(entity.getDescription());
             resource.setName(entity.getName());
+            resource.setOwningTenant(entity.getOwningTenant().getTenantId());
 
             return resource;
         }
